requirements.txt
pandas
numpy
scikit-learn
streamlit
sqlalchemy
shap
xgboost
lightgbm
imbalanced-learn
python-dotenv
requests
joblib
